move json encoder to base and use in l2cache
