+++ caveclient/base.py
@@ -3,6 +3,18 @@ import requests
 import json
 import logging
 import webbrowser
+import numpy as np
+import datetime
+
+class CGEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()
+        if isinstance(obj, np.uint64):
+            return int(obj)
+        if isinstance(obj, (datetime.datetime, datetime.date)):
+            return obj.isoformat()
+        return json.JSONEncoder.default(self, obj)
 
 
 class AuthException(Exception):
